refactor(sqlite): extract CLI subcommands into modular cli/ subpackage
Split the monolithic __main__.py (~1335 lines) into focused subcommand
modules under cli/:

- __init__.py: subcommand registration and dynamic handler dispatch
- output.py: output provider factory and display helpers
- connection.py: connection argument parsing and backend creation
- info.py: SQLite version/protocol info (no connection args needed)
- query.py: SQL execution with --log-level
- introspect.py: database introspection
- status.py: server status overview
- named_query.py: named-query adapter for shared CLI helper
- named_procedure.py: named-procedure adapter for shared CLI helper
- named_connection.py: named-connection adapter (no -o, no connection args)

Each subcommand now independently decides its own arguments:
- info: only -o {table,json}, --verbose, --rich-ascii (no --db-file)
- query: all connection args, -o {table,json,csv,tsv}, --log-level
- introspect/status: all connection args, -o {table,json,csv,tsv}
- named-query/named-procedure: connection parent parser, -o, --rich-ascii
- named-connection: empty parent parser (no -o, no connection args)
