+++ src/rhosocial/activerecord/backend/impl/sqlite/cli/info.py
@@ -1,0 +1,418 @@
+# src/rhosocial/activerecord/backend/impl/sqlite/cli/info.py
+"""info subcommand - Display SQLite environment information.
+
+The info command shows information derived entirely from the Python sqlite3
+module version, independent of any database file, so it does not need
+--db-file or other connection arguments.
+"""
+
+import argparse
+import inspect
+import json
+import sys
+from typing import Dict, List, Any
+
+from rhosocial.activerecord.backend.impl.sqlite.extension import get_registry
+from rhosocial.activerecord.backend.impl.sqlite.pragma import get_all_pragma_infos, PragmaCategory
+from rhosocial.activerecord.backend.impl.sqlite.protocols import SQLiteExtensionSupport, SQLitePragmaSupport
+from rhosocial.activerecord.backend.dialect.protocols import (
+    WindowFunctionSupport,
+    CTESupport,
+    ReturningSupport,
+    UpsertSupport,
+    LateralJoinSupport,
+    JoinSupport,
+    JSONSupport,
+    ExplainSupport,
+    GraphSupport,
+    FilterClauseSupport,
+    SetOperationSupport,
+    ViewSupport,
+    TableSupport,
+    TruncateSupport,
+    GeneratedColumnSupport,
+    TriggerSupport,
+    FunctionSupport,
+    AdvancedGroupingSupport,
+    ArraySupport,
+    ILIKESupport,
+    IndexSupport,
+    LockingSupport,
+    MergeSupport,
+    OrderedSetAggregationSupport,
+    QualifyClauseSupport,
+    SchemaSupport,
+    SequenceSupport,
+    TemporalTableSupport,
+)
+
+from .connection import create_backend_from_memory
+from .output import RICH_AVAILABLE
+
+# info only supports table and json output (csv/tsv not suitable for nested structures)
+OUTPUT_CHOICES = ['table', 'json']
+
+# Groups that are specific to SQLite dialect
+DIALECT_SPECIFIC_GROUPS = {"SQLite-specific"}
+
+PROTOCOL_FAMILY_GROUPS: Dict[str, list] = {
+    "Query Features": [
+        WindowFunctionSupport,
+        CTESupport,
+        FilterClauseSupport,
+        SetOperationSupport,
+        AdvancedGroupingSupport,
+    ],
+    "JOIN Support": [JoinSupport, LateralJoinSupport],
+    "Data Types": [JSONSupport, ArraySupport],
+    "DML Features": [
+        ReturningSupport,
+        UpsertSupport,
+        MergeSupport,
+        OrderedSetAggregationSupport,
+    ],
+    "Transaction & Locking": [LockingSupport, TemporalTableSupport],
+    "Query Analysis": [ExplainSupport, GraphSupport, QualifyClauseSupport],
+    "DDL - Table": [TableSupport, TruncateSupport, GeneratedColumnSupport],
+    "DDL - View": [ViewSupport],
+    "DDL - Schema & Index": [SchemaSupport, IndexSupport],
+    "DDL - Sequence & Trigger": [SequenceSupport, TriggerSupport, FunctionSupport],
+    "String Matching": [ILIKESupport],
+    "SQLite-specific": [
+        SQLiteExtensionSupport,
+        SQLitePragmaSupport,
+    ],
+}
+
+# All possible test arguments for methods that require parameters
+SUPPORT_METHOD_ALL_ARGS: Dict[str, List[str]] = {
+    "supports_explain_format": ["TEXT", "JSON", "XML", "YAML", "TREE", "DOT"],
+    "is_extension_available": [
+        "fts5", "fts4", "fts3", "fts2", "fts1",
+        "json1", "rtree", "geopoly", "dbstat", "fts5tokenize",
+    ],
+    "is_pragma_available": [
+        "journal_mode", "synchronous", "cache_size",
+        "temp_store", "foreign_keys", "busy_timeout", "wal_autocheckpoint",
+    ],
+}
+
+
+def create_parser(subparsers):
+    """Create the info subcommand parser."""
+    parser = subparsers.add_parser(
+        'info',
+        help='Display SQLite environment information (uses in-memory database)',
+        epilog="""Examples:
+  # Show basic info (in-memory database)
+  %(prog)s info
+
+  # Show detailed protocol support
+  %(prog)s info -vv
+
+  # Show info as JSON
+  %(prog)s info -o json
+""",
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+    )
+
+    # Output format (table and json only; csv/tsv not suitable for nested structures)
+    parser.add_argument(
+        '-o', '--output',
+        choices=OUTPUT_CHOICES,
+        default='table',
+        help='Output format (default: table)',
+    )
+
+    # Verbosity
+    parser.add_argument(
+        '-v', '--verbose',
+        action='count',
+        default=0,
+        help='Increase verbosity. -v for families, -vv for details.',
+    )
+
+    # Rich display options
+    parser.add_argument(
+        '--rich-ascii',
+        action='store_true',
+        help='Use ASCII characters for rich table borders.',
+    )
+
+    return parser
+
+
+def handle(args):
+    """Handle the info subcommand.
+
+    info uses an in-memory database; displayed information is determined
+    entirely by the Python sqlite3 module version.
+    """
+    backend = create_backend_from_memory()
+
+    output_format = args.output if args.output != "table" or RICH_AVAILABLE else "json"
+
+    try:
+        dialect = backend.dialect
+        version_tuple = dialect.version
+        sqlite_version = ".".join(map(str, version_tuple))
+    except Exception as e:
+        if output_format == "json" or not RICH_AVAILABLE:
+            print(json.dumps({"error": f"Failed to connect: {e}"}))
+        else:
+            print(f"Error: Failed to connect to database: {e}")
+        return None
+    finally:
+        backend.disconnect()
+
+    # Build info structure using helper functions
+    info = {
+        "database": _build_database_info(version_tuple),
+        "features": {
+            "extensions": _build_extension_info(version_tuple),
+            "pragmas": _build_pragma_info(),
+        },
+        "protocols": _build_protocol_info(dialect, args.verbose),
+    }
+
+    # Add connection info (always memory for info)
+    info["database"]["database_type"] = "memory"
+    info["database"]["database_path"] = ":memory:"
+
+    # Output result
+    if output_format == "json" or not RICH_AVAILABLE:
+        print(json.dumps(info, indent=2))
+    else:
+        info_legacy = {
+            "sqlite": info["database"],
+            "extensions": info["features"]["extensions"],
+            "pragmas": info["features"]["pragmas"],
+            "protocols": info["protocols"],
+        }
+        _display_info_rich(info_legacy, args.verbose, sqlite_version, False, ":memory:")
+
+    return info
+
+
+# ---------------------------------------------------------------------------
+# Internal helper functions
+# ---------------------------------------------------------------------------
+
+def _get_protocol_support_methods(protocol_class: type) -> List[str]:
+    """Get all support check methods from a protocol class."""
+    methods = []
+    for name, member in inspect.getmembers(protocol_class):
+        if callable(member) and (
+            name.startswith("supports_") or name.startswith("is_") and name.endswith("_available")
+        ):
+            methods.append(name)
+    return sorted(methods)
+
+
+def _check_protocol_support(dialect: Any, protocol_class: type) -> Dict[str, Any]:
+    """Check all support methods for a protocol against the dialect."""
+    results = {}
+    methods = _get_protocol_support_methods(protocol_class)
+    for method_name in methods:
+        if hasattr(dialect, method_name):
+            try:
+                method = getattr(dialect, method_name)
+                sig = inspect.signature(method)
+                params = [p for p in sig.parameters.values() if p.default == inspect.Parameter.empty]
+                required_params = [p for p in params if p.name != "self"]
+
+                if len(required_params) == 0:
+                    result = method()
+                    results[method_name] = bool(result)
+                elif method_name in SUPPORT_METHOD_ALL_ARGS:
+                    all_args = SUPPORT_METHOD_ALL_ARGS[method_name]
+                    arg_results = {}
+                    for arg in all_args:
+                        try:
+                            arg_results[arg] = bool(method(arg))
+                        except Exception:
+                            arg_results[arg] = False
+                    supported_count = sum(1 for v in arg_results.values() if v)
+                    results[method_name] = {"supported": supported_count, "total": len(all_args), "args": arg_results}
+                else:
+                    results[method_name] = False
+            except Exception:
+                results[method_name] = False
+        else:
+            results[method_name] = False
+    return results
+
+
+def _build_database_info(version_tuple: tuple) -> Dict[str, Any]:
+    """Build database basic information structure."""
+    return {
+        "type": "sqlite",
+        "version": ".".join(map(str, version_tuple)),
+        "version_tuple": list(version_tuple),
+    }
+
+
+def _build_extension_info(version_tuple: tuple) -> Dict[str, Any]:
+    """Build extension information structure."""
+    registry = get_registry()
+    extensions = registry.detect_extensions(version_tuple)
+    ext_info = {}
+
+    for name, ext in extensions.items():
+        ext_info[name] = {
+            "type": ext.extension_type.name,
+            "available": ext.installed,
+            "min_version": ".".join(map(str, ext.min_version)),
+            "deprecated": ext.deprecated,
+            "description": ext.description,
+        }
+        if ext.successor:
+            ext_info[name]["successor"] = ext.successor
+        if ext.installed:
+            features = registry.get_supported_features(name, version_tuple)
+            if features:
+                ext_info[name]["features"] = features
+
+    return ext_info
+
+
+def _build_pragma_info() -> Dict[str, Any]:
+    """Build pragma information structure."""
+    pragma_info = {"total_count": len(get_all_pragma_infos()), "categories": {}}
+
+    for category in PragmaCategory:
+        pragmas_in_category = [name for name, p in get_all_pragma_infos().items() if p.category == category]
+        pragma_info["categories"][category.name] = {"count": len(pragmas_in_category), "names": pragmas_in_category}
+
+    return pragma_info
+
+
+def _calculate_support_stats(support_methods: Dict[str, Any]) -> tuple:
+    """Calculate supported/total counts from support methods."""
+    supported_count = 0
+    total_count = 0
+
+    for value in support_methods.values():
+        if isinstance(value, dict):
+            supported_count += value["supported"]
+            total_count += value["total"]
+        else:
+            total_count += 1
+            if value:
+                supported_count += 1
+
+    return supported_count, total_count
+
+
+def _build_protocol_info(dialect: Any, verbose: int) -> Dict[str, Any]:
+    """Build protocol support information structure."""
+    protocol_info = {}
+    for group_name, protocols in PROTOCOL_FAMILY_GROUPS.items():
+        protocol_info[group_name] = _build_protocol_group_info(dialect, protocols, verbose)
+    return protocol_info
+
+
+def _build_protocol_group_info(dialect: Any, protocols: list, verbose: int) -> Dict[str, Any]:
+    """Build information for a single protocol group."""
+    group_info = {}
+
+    for protocol in protocols:
+        protocol_name = protocol.__name__
+        support_methods = _check_protocol_support(dialect, protocol)
+        supported_count, total_count = _calculate_support_stats(support_methods)
+
+        percentage = round(supported_count / total_count * 100, 1) if total_count > 0 else 0
+
+        group_info[protocol_name] = {
+            "supported": supported_count,
+            "total": total_count,
+            "percentage": percentage,
+        }
+
+        if verbose >= 2:
+            group_info[protocol_name]["methods"] = support_methods
+
+    return group_info
+
+
+def _display_info_rich(
+    info: Dict, verbose: int, sqlite_version: str,
+    is_file_database: bool = True, db_path: str = ":memory:"
+):
+    """Display info using rich console."""
+    from rich.console import Console
+
+    console = Console(force_terminal=True)
+
+    SYM_OK = "[OK]"
+    SYM_PARTIAL = "[~]"
+    SYM_FAIL = "[X]"
+
+    console.print("\n[bold cyan]SQLite Environment Information[/bold cyan]\n")
+
+    if is_file_database:
+        console.print(f"[bold]SQLite Version:[/bold] {sqlite_version} [dim](file: {db_path})[/dim]\n")
+    else:
+        console.print(f"[bold]SQLite Version:[/bold] {sqlite_version} [yellow](in-memory database)[/yellow]\n")
+
+    console.print("[bold green]Extension Support:[/bold green]")
+    for name, ext in info["extensions"].items():
+        status = "[green][OK][/green]" if ext["available"] else "[red][X][/red]"
+        deprecated = " [yellow](deprecated)[/yellow]" if ext["deprecated"] else ""
+        console.print(f"  {status} [bold]{name}[/bold]{deprecated}: {ext['description']}")
+        if ext["available"] and "features" in ext:
+            console.print(f"      [dim]Features: {', '.join(ext['features'])}[/dim]")
+
+    console.print("\n[bold green]Pragma System:[/bold green]")
+    console.print(f"  Total pragmas documented: {info['pragmas']['total_count']}")
+    for cat_name, cat_info in info["pragmas"]["categories"].items():
+        console.print(f"  [bold]{cat_name}:[/bold] {cat_info['count']} pragmas")
+
+    label = "Detailed" if verbose >= 2 else "Family Overview"
+    console.print(f"\n[bold green]Protocol Support ({label}):[/bold green]")
+
+    for group_name, protocols in info["protocols"].items():
+        if group_name in DIALECT_SPECIFIC_GROUPS:
+            console.print(f"\n  [bold underline]{group_name}:[/bold underline] [dim](dialect-specific)[/dim]")
+        else:
+            console.print(f"\n  [bold underline]{group_name}:[/bold underline]")
+        for protocol_name, stats in protocols.items():
+            pct = stats["percentage"]
+            if pct == 100:
+                color = "green"
+                symbol = SYM_OK
+            elif pct >= 50:
+                color = "yellow"
+                symbol = SYM_PARTIAL
+            elif pct > 0:
+                color = "red"
+                symbol = SYM_PARTIAL
+            else:
+                color = "red"
+                symbol = SYM_FAIL
+
+            bar_len = 20
+            filled = int(pct / 100 * bar_len)
+            bar = "#" * filled + "-" * (bar_len - filled)
+
+            sup = stats["supported"]
+            tot = stats["total"]
+            console.print(
+                f"    [{color}]{symbol}[/{color}] {protocol_name}: [{color}]{bar}[/{color}] {pct:.0f}% ({sup}/{tot})"
+            )
+
+            if verbose >= 2 and "methods" in stats:
+                for method, value in stats["methods"].items():
+                    method_display = (
+                        method.replace("supports_", "").replace("_", " ").replace("is_", "").replace("_available", "")
+                    )
+                    if isinstance(value, dict):
+                        console.print(f"        [dim]{method_display}:[/dim]")
+                        for arg, supported in value.get("args", {}).items():
+                            m_status = "[green][OK][/green]" if supported else "[red][X][/red]"
+                            console.print(f"            {m_status} {arg}")
+                    else:
+                        m_status = "[green][OK][/green]" if value else "[red][X][/red]"
+                        console.print(f"        {m_status} {method_display}")
+
+    console.print()
